Clean code: Move DirectionMapper to the mapping package.
